P value is now calculated from Poisson test
Previously, the p-value was determined from a Chi-square test of the observed and expected frequencies from each of the ten years of data. Now that the model uses continuous time it is instead calculated from a test of two Poisson distributions with the rates equal to the number of prevalent cases from both the known value and the simulated value. This has necessitated making the simulation run over the years that we have registry data for, despite not needing these prevalence estimates for the overall estimation since they are replaced with the counted numbers.