Moved guards to main function.
This leaves sim_prevalence as a function that must be correctly specified and isn't designed to be regularly used directly by users, such as survreg.fit, however it can be done to speed code up.

However, the bootstrapping code currently doesn't work as intended, since the model formula used to build the model is saved in a variable unique to the environment of the calling function, and so when the call is evaluated in a different function/environment it cannot find this formula. I've posted on Stack Overflow to try and find a way of still using a variable but allowing the exact variables to be saved in the call. https://stackoverflow.com/questions/44849941/r-saving-function-call-with-formula-for-reuse-in-bootstrapping

As a result, the code currently isn't in a stable working version, but as I'm on holiday for a week I want to ensure it's backed up.