sim_prevalence now only accepts objects.
This allows for the optimisation of sim_prevalence as a function that accepts fully specified survival and incidence objects, whereas the higher level prevalence function can also accept a formula and will form default objects if possible. sim_prevalence can still be made accessible to the user although as with the survreg.fit functions, it isn't intended for everyday use.