Incident date column is no longer hardcoded.
The entry date column itself doesn't need to be specified for the code to work, as long as the registry_start_date is provided, so that the appropriate simulation time can be determined. If the column holding incident dates is unknown then the prevalence estimates are simply solely derived from simulation, while they are supplemented with counted estimates if incident date and death date columns are provided.