Handles flexsurvcure objects
A predict_survival_probability() method for flexsurvcure objects has been provided, along with a constructor wrapper that allows the use of population mortality rates to scale survival estimates. In the current flexsurvcure implementation it is assumed that the survival probability of cured individuals is always 1. The wrapper allows population mortality tables to be passed in instead.